fixed Api, might make it so less points show, or maybe make it so that each point isn't fully visible
